#3 design : LargeModal 공통컴포넌트  구현
