+++ components/Common/Modal/index.ts
@@ -1,0 +1,1 @@
+export { default as LargeModal } from './LargeModal';
