+++ components/Home/index.tsx
@@ -1,1 +1,2 @@
 export { default as Home } from './Home';
+export { default as Search } from './Search';
