+++ components/RoomList/index.ts
@@ -1,3 +1,4 @@
 export { default as AddStudyroom } from './AddStudyroom';
+export { default as AddStudyroomModal } from './AddStudyroomModal';
 export { default as StudyroomCard } from './StudyroomCard';
 export { default as StudyroomList } from './StudyroomList';
